+++ src/stylesheets/style.css
@@ -7,7 +7,6 @@
 @import 'superModal.css';
 @import 'snackbar.css';
 
-
 :root {
   --bg: transparent;
   --onBg: transparent;
